refactor: Move webidentity creation code into aws package
