+++ pkg/controllers/kluctl_project.go
@@ -18,7 +18,6 @@ import (
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"os"
 	"path/filepath"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 	"strings"
 	"time"
 
@@ -547,13 +546,8 @@ func (pp *preparedProject) addServiceAccountBasedKeyServers(ctx context.Context,
 	if name == "" {
 		return nil
 	}
-	var sa corev1.ServiceAccount
-	err := pp.r.Client.Get(ctx, client.ObjectKey{Name: name, Namespace: pp.obj.Namespace}, &sa)
-	if err != nil {
-		return fmt.Errorf("failed to retrieve service account %s: %w", name, err)
-	}
 
-	ks, err := sops.BuildSopsKeyServerFromServiceAccount(ctx, pp.r.Client, &sa)
+	ks, err := sops.BuildSopsKeyServerFromServiceAccount(ctx, pp.r.Client, name, pp.obj.Namespace)
 	if err != nil {
 		return err
 	}
